refactor: java_convert_kotlin 패키지 이름 변경 -> study
