+++ src/main/kotlin/study/Person.kt
@@ -1,3 +1,3 @@
-package java_convert_kotlin
+package study
 
 data class Person(val name: String, val age: Int?, var nickname: String? = "찰리")
